Vue課題②-1修正(1回目)
