+++ src/main.js
@@ -2,14 +2,14 @@ import Vue from 'vue'
 import App from './App.vue'
 import router from './router'
 import store from './store'
+
 import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
 import { BootstrapVue, IconsPlugin } from 'bootstrap-vue'
-
-Vue.config.productionTip = false
-
 Vue.use(BootstrapVue)
 Vue.use(IconsPlugin)
+
+Vue.config.productionTip = false
 
 new Vue({
   router,
